[TECH-TASK][FIRST-TECH-TASK] Code improvement

+++ src/main/java/com/cenfo/tech/task1/services/category/CategoryService.java
@@ -3,13 +3,13 @@ package com.cenfo.tech.task1.services.category;
 import com.cenfo.tech.task1.entity.Category;
 import com.cenfo.tech.task1.entity.Product;
 import com.cenfo.tech.task1.repository.ICategoryRepository;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import com.cenfo.tech.task1.response.dto.CategoryDTO;
+import com.cenfo.tech.task1.utils.UtilsDTO;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Collections;
-import java.util.List;
 
 @Service
 public class CategoryService implements ICategoryService {
@@ -22,44 +22,52 @@ public class CategoryService implements ICategoryService {
 
     @Transactional
     @Override
-    public ResponseEntity<Category> register(Category category) {
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(categoryRepository.save(category));
+    public CategoryDTO register(Category category) {
+        try {
+            return UtilsDTO.toCategoryDTO(categoryRepository.save(category));
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
 
     @Override
-    public ResponseEntity<List<Category>> getAll() {
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(categoryRepository.findAll());
+    public Page<Category> getAll(int page, int size) {
+        Page<Category> categoryPage = categoryRepository.findAll(PageRequest.of(page, size));
+        if (categoryPage.isEmpty()) {
+            throw new EntityNotFoundException("No categories found on page " + page + " with size " + size);
+        }
+        return categoryPage;
     }
 
     @Override
-    public ResponseEntity<Category> getById(Long id) {
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(categoryRepository.findById(id).orElse(null));
-    }
-
-    @Override
-    public ResponseEntity<List<Product>> getAllProductsByCategory(Long id) {
+    public CategoryDTO getById(Long id) {
         return categoryRepository.findById(id)
-                .map(category -> ResponseEntity.ok(category.getProducts()))
-                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).body(Collections.emptyList()));
+                .map(UtilsDTO::toCategoryDTO)
+                .orElseThrow(() -> new EntityNotFoundException("Category not found with id: " + id));
+    }
+
+    @Override
+    public Page<Product> getAllProductsByCategory(Long categoryId, int page, int size) {
+        Page<Product> productPage = categoryRepository.findProductsByCategoryId(categoryId, PageRequest.of(page, size));
+        if (productPage.isEmpty()) {
+            throw new EntityNotFoundException("No products found for category " + categoryId + " on page " + page + " with size " + size);
+        }
+        return productPage;
     }
 
     @Transactional
     @Override
-    public ResponseEntity<Category> update(Long id, Category category) {
-        return categoryRepository.findById(id)
-                .map(existingCategory -> {
-                    category.setId(id);
-                    return ResponseEntity.ok(categoryRepository.save(category));
-                })
-                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+    public CategoryDTO update(Long id, Category category) {
+        if (categoryRepository.findById(id).isPresent()) {
+            category.setId(id);
+            return UtilsDTO.toCategoryDTO(categoryRepository.save(category));
+        } else throw new EntityNotFoundException("Category not found with id: " + id);
     }
-
 
     @Override
     public void delete(Long id) {
-        categoryRepository.deleteById(id);
+        if (categoryRepository.findById(id).isPresent()) {
+            categoryRepository.deleteById(id);
+        } else throw new EntityNotFoundException("Category not found with id: " + id);
     }
 }
